requirements.txt
boto3==1.37.5
pydantic-settings==2.8.1
openai==1.65.5
textract-trp==0.1.3
pydantic==2.10.6
Pyrogram==2.0.106
TgCrypto==1.2.5
